管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -47,6 +47,9 @@ public class EmployeeAction extends ActionBase {
      */
     public void index() throws ServletException, IOException {
 
+        //管理者かどうかのチェック　//追記
+        if(checkAdmin()) { //追記
+
         //指定されたページ数の一覧画面に表示するデータを取得
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
@@ -68,7 +71,9 @@ public class EmployeeAction extends ActionBase {
 
         //一覧画面を表示
         forward(ForwardConst.FW_EMP_INDEX);
+        }//追記
     }
+
 
 
     //-以下、6.6Employeeに対するアクションとビューの作成１：indexの作成----------------------------------------------------------------
@@ -83,11 +88,14 @@ public class EmployeeAction extends ActionBase {
 
         public void entryNew() throws ServletException, IOException {
 
+            //管理者か同課のチェック　//追記
+            if (checkAdmin()) { //追記
             putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
             putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
 
             //新規登録画面を表示
             forward(ForwardConst.FW_EMP_NEW);
+            }//追記
         }
 
       //--(6.8 create）----------------------------------------------------------------
@@ -99,7 +107,7 @@ public class EmployeeAction extends ActionBase {
         public void create() throws ServletException, IOException {
 
             //CSRF対策 tokenのチェック
-            if (checkToken()) {
+            if (checkAdmin()&&checkToken()) {//追記
 
                 //パラメータの値を元に従業員情報のインスタンスを作成する
                 EmployeeView ev = new EmployeeView(
@@ -150,6 +158,8 @@ public class EmployeeAction extends ActionBase {
          */
         public void show() throws ServletException, IOException {
 
+            //管理者かどうかのチェック//追記
+            if (checkAdmin()) {//追記
             //idを条件に従業員データを取得する
             EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -164,6 +174,7 @@ public class EmployeeAction extends ActionBase {
 
             //詳細画面を表示
             forward(ForwardConst.FW_EMP_SHOW);
+            }//追記
         }
 
         //--(6.10 editメソッド）----------------------------------------------------------------
@@ -175,6 +186,9 @@ public class EmployeeAction extends ActionBase {
      */
 
     public void edit() throws ServletException,IOException {
+
+        //管理者かどうかのチェック//追記
+        if(checkAdmin()) {
 
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -191,9 +205,10 @@ public class EmployeeAction extends ActionBase {
 
         //編集画面を表示する
         forward(ForwardConst.FW_EMP_EDIT);
+        }//追記
 
+    }
 
-        }
     //--(6.11 updateメソッド）----------------------------------------------------------------
     /**
     * 更新を行う
@@ -201,10 +216,10 @@ public class EmployeeAction extends ActionBase {
     * @throws IOException
     */
 
-    public void update() throws ServletException,IOException {
+    public void update() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        if (checkAdmin()&&checkToken()) {//追記
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -253,19 +268,47 @@ public class EmployeeAction extends ActionBase {
     public void destroy() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        if (checkAdmin()&&checkToken()) {//追記
 
             //idを条件に従業員データを論理削除する
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
             //セッションに削除完了のフラッシュメッセージを設定
-            putSessionScope(AttributeConst.FLUSH,MessageConst.I_DELETED.getMessage());
+            putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
 
             //一覧画面にリダイレクト
-            redirect(ForwardConst.ACT_EMP,ForwardConst.CMD_INDEX);
+            redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
         }
     }
+
+
+
+/**
+ * ログイン中の従業員が管理者かどうかチェックし、管理者でなければエラー画面を表示
+ * true: 管理者 false: 管理者ではない
+ * @throws ServletException
+ * @throws IOException
+ */
+private boolean checkAdmin() throws ServletException, IOException {
+
+    //セッションからログイン中の従業員情報を取得
+    EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+    //管理者でなければエラー画面を表示
+    if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+        forward(ForwardConst.FW_ERR_UNKNOWN);
+        return false;
+
+    } else {
+
+        return true;
+    }
+
 }
+
+}
+
